+++ src/pages/dashboard/pages/Products.jsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
 import { useLoaderData } from "react-router-dom";
-import { IconButtonWithBgGold } from "src/components/buttons";
 import { TableRow } from "./index";
 import { Loader } from "src/components";
 import axios from "axios";
@@ -48,7 +47,7 @@ export function Products() {
         <thead className="border border-gold">
           <tr className="goldGradient text-base uppercase">
             {colsTitles.map((t, index) => (
-              <th className="border-r-blue px-6 py-2 border-r-2 last:border-none" key={index}>
+              <th className="border-r-2 border-r-blue px-2  py-2 text-xs last:border-none" key={index}>
                 {t}
               </th>
             ))}
@@ -70,7 +69,7 @@ export function Products() {
               <TableRow
                 content={
                   <i
-                    className={`ri-${published ? "check" : "close"}-fill text-3xl text-${
+                    className={`ri-${published ? "check" : "close"}-fill text-2xl text-${
                       published ? "green" : "red"
                     }-500`}
                   />
@@ -78,7 +77,7 @@ export function Products() {
               />
               <TableRow
                 content={
-                  <label htmlFor="fileInput" className="icons ri-image-2-fill text-3xl">
+                  <label htmlFor="fileInput" className="icons ri-image-2-fill text-2xl">
                     <input
                       type="file"
                       id="fileInput"
@@ -93,11 +92,6 @@ export function Products() {
           ))}
         </tbody>
       </table>
-      <IconButtonWithBgGold
-        className={"absolute bottom-0 mx-auto my-4"}
-        icon={"ri-arrow-up-s-line"}
-        onClick={() => window.scrollTo({ top: 0, behavior: "smooth" })}
-      />
     </main>
   );
 }
