+++ src/pages/dashboard/pages/Users.jsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
 import { useLoaderData } from "react-router-dom";
-import { IconButtonWithBgGold } from "src/components/buttons";
 import { TableRow } from "./index";
 import { APIHydro } from "src/api";
 import { Loader } from "src/components";
@@ -52,7 +51,7 @@ export function Users() {
         <thead className="border border-gold">
           <tr className="goldGradient text-base uppercase">
             {colsTitles.map((t, index) => (
-              <th className="border-r-2 border-r-blue px-6 py-2 last:border-none" key={index}>
+              <th className="border-r-2 border-r-blue px-2 py-2 text-xs last:border-none" key={index}>
                 {t}
               </th>
             ))}
@@ -69,7 +68,7 @@ export function Users() {
                   <i
                     className={`icons ri-user-${
                       type === "ADMIN" ? "star-line textGoldGradient" : "line text-white"
-                    } text-3xl`}
+                    } text-2xl`}
                     onClick={() => handleAdmin(id, name, type)}
                   />
                 }
@@ -78,7 +77,7 @@ export function Users() {
               <TableRow
                 content={
                   <i
-                    className={`ri-${active ? "check" : "close"}-fill text-3xl ${
+                    className={`ri-${active ? "check" : "close"}-fill text-2xl ${
                       active ? "text-green-500" : "text-red-500"
                     }`}
                   />
@@ -88,11 +87,6 @@ export function Users() {
           ))}
         </tbody>
       </table>
-      <IconButtonWithBgGold
-        className={"absolute bottom-0 mx-auto my-4"}
-        icon={"ri-arrow-up-s-line"}
-        onClick={() => window.scrollTo({ top: 0, behavior: "smooth" })}
-      />
     </main>
   );
 }
