+++ src/pages/dashboard/pages/Utils.jsx
@@ -1,3 +1,3 @@
 export function TableRow({ content, style = "text-center" }) {
-  return <td className={`border border-gold px-4 py-2 ${style}`}>{content}</td>;
+  return <td className={`border border-gold px-4 py-2 text-xs ${style}`}>{content}</td>;
 }
